manager: Ensure that at least one string is set when creating gadget
Kernel driver provides set of strings (manufacturer, serial,
product) to USB host only if at least one of them has been
set. In gadgetd we follow the most popular convention that
each USB device should provide strings in English (US).

If user didn't provide value for any of strings we create
empty one. This is information for driver that this gadget
provides strings in English and all three are empty.

Change-Id: I992dc38cebe37b34431b40d91e9f24a41062daf7
Signed-off-by: Karol Lewandowski <k.lewandowsk@samsung.com>
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>