gadgetd: Remove use of usbg_function_attrs type
In libusbgx, usbg_function_attrs type has gone away and in the example
given in libusbgx, replaced with a union.  This seems like a horrible
way to implement this from a user of the library point of view and
further refactoring of this code in gadgetd should take place to correct
this.

Signed-off-by: Andrew Bradford <andrew.bradford@kodakalaris.com>